normd calculations are in double precision, only the result is casted back to single precision

+++ include/norm.h
@@ -5,8 +5,8 @@
 #include "config.h"
 
 
-float  normf(const float x[], int n);
-double normd(const float x[], int n);
+float normf(const float x[], int n);
+float normd(const float x[], int n);
 
 
 #endif
